requirements.txt
openai
langchain
langchain_community
autogen
pymupdf
langchain_openai
tiktoken
python-dotenv
streamlit

# For local LLM (Llama 3 via Ollama) and HuggingFace Embeddings
sentence-transformers
langchain_community
ollama-python

flask
flask-cors
flask-jwt-extended
python-dotenv
fitz
PyMuPDF
requests
google-generativeai
autogen
langchain-community
chromadb
duckduckgo-search
serpapi

